+++ src/emulator/functionsEmulator.ts
@@ -33,6 +33,7 @@ import { EventEmitter } from "events";
 import * as stream from "stream";
 import { EmulatorLogger, Verbosity } from "./emulatorLogger";
 import { RuntimeWorkerPool, RuntimeWorker } from "./functionsRuntimeWorker";
+import { PubsubEmulator } from "./pubsubEmulator";
 import { FirebaseError } from "../error";
 import { WorkQueue } from "./workQueue";
 
@@ -188,6 +189,7 @@ export class FunctionsEmulator implements EmulatorInstance {
       ports: {
         firestore: EmulatorRegistry.getPort(Emulators.FIRESTORE),
         database: EmulatorRegistry.getPort(Emulators.DATABASE),
+        pubsub: EmulatorRegistry.getPort(Emulators.PUBSUB),
       },
       proto,
       triggerId,
@@ -293,6 +295,9 @@ export class FunctionsEmulator implements EmulatorInstance {
             case Constants.SERVICE_REALTIME_DATABASE:
               added = await this.addRealtimeDatabaseTrigger(this.args.projectId, definition);
               break;
+            case Constants.SERVICE_PUBSUB:
+              added = await this.addPubsubTrigger(this.args.projectId, definition);
+              break;
             default:
               EmulatorLogger.log("DEBUG", `Unsupported trigger: ${JSON.stringify(definition)}`);
               break;
@@ -344,7 +349,7 @@ export class FunctionsEmulator implements EmulatorInstance {
     if (!databasePort) {
       return Promise.resolve(false);
     }
-    if (definition.eventTrigger === undefined) {
+    if (!definition.eventTrigger) {
       EmulatorLogger.log(
         "WARN",
         `Event trigger "${definition.name}" has undefined "eventTrigger" member`
@@ -433,6 +438,36 @@ export class FunctionsEmulator implements EmulatorInstance {
     });
   }
 
+  async addPubsubTrigger(
+    projectId: string,
+    definition: EmulatedTriggerDefinition
+  ): Promise<boolean> {
+    const pubsubPort = EmulatorRegistry.getPort(Emulators.PUBSUB);
+    if (!pubsubPort) {
+      return false;
+    }
+
+    if (!definition.eventTrigger) {
+      return false;
+    }
+
+    const pubsubEmulator = EmulatorRegistry.get(Emulators.PUBSUB) as PubsubEmulator;
+
+    logger.debug(`addPubsubTrigger`, JSON.stringify({ eventTrigger: definition.eventTrigger }));
+
+    // "resource":\"projects/{PROJECT_ID}/topics/{TOPIC_ID}";
+    const resource = definition.eventTrigger.resource;
+    const resourceParts = resource.split("/");
+    const topic = resourceParts[resourceParts.length - 1];
+
+    try {
+      await pubsubEmulator.addTrigger(topic, definition.name);
+      return true;
+    } catch (e) {
+      return false;
+    }
+  }
+
   getProjectId(): string {
     return this.args.projectId;
   }
@@ -474,6 +509,7 @@ export class FunctionsEmulator implements EmulatorInstance {
       ports: {
         firestore: EmulatorRegistry.getPort(Emulators.FIRESTORE),
         database: EmulatorRegistry.getPort(Emulators.DATABASE),
+        pubsub: EmulatorRegistry.getPort(Emulators.PUBSUB),
       },
       disabled_features: this.args.disabledRuntimeFeatures,
     };