+++ src/emulator/functionsEmulatorRuntime.ts
@@ -744,6 +744,12 @@ function initializeEnvironmentalVariables(frb: FunctionsRuntimeBundle): void {
       }
     }
   }
+
+  if (frb.ports.pubsub && isFeatureEnabled(frb, "pubsub_emulator")) {
+    const pubsubHost = `localhost:${frb.ports.pubsub}`;
+    process.env.PUBSUB_EMULATOR_HOST = pubsubHost;
+    logDebug(`Set PUBSUB_EMULATOR_HOST to ${pubsubHost}`);
+  }
 }
 
 async function initializeFunctionsConfigHelper(functionsDir: string): Promise<void> {
@@ -803,11 +809,16 @@ async function processHTTPS(frb: FunctionsRuntimeBundle, trigger: EmulatedTrigge
   await new Promise((resolveEphemeralServer, rejectEphemeralServer) => {
     const handler = async (req: express.Request, res: express.Response) => {
       try {
-        logDebug(`Ephemeral server used!`);
+        logDebug(`Ephemeral server handling ${req.method} request`);
         const func = trigger.getRawFunction();
-
         res.on("finish", () => {
-          instance.close(resolveEphemeralServer);
+          instance.close((err) => {
+            if (err) {
+              rejectEphemeralServer(err);
+            } else {
+              resolveEphemeralServer();
+            }
+          });
         });
 
         await runHTTPS([req, res], func);
@@ -851,9 +862,12 @@ async function processHTTPS(frb: FunctionsRuntimeBundle, trigger: EmulatedTrigge
       functionRouter
     );
 
+    logDebug(`Attempting to listen to socketPath: ${socketPath}`);
     const instance = ephemeralServer.listen(socketPath, () => {
       new EmulatorLog("SYSTEM", "runtime-status", "ready", { state: "ready" }).log();
     });
+
+    instance.on("error", rejectEphemeralServer);
   });
 }
 
@@ -884,20 +898,12 @@ async function processBackground(
  * Run the given function while redirecting logs and looking out for errors.
  */
 async function runFunction(func: () => Promise<any>): Promise<any> {
-  /* tslint:disable:no-console */
-  const log = console.log;
-  console.log = (...messages: any[]) => {
-    new EmulatorLog("USER", "function-log", messages.join(" ")).log();
-  };
-
   let caughtErr;
   try {
     await func();
   } catch (err) {
     caughtErr = err;
   }
-
-  console.log = log;
 
   logDebug(`Ephemeral server survived.`);
   if (caughtErr) {
@@ -957,7 +963,7 @@ async function moduleResolutionDetective(frb: FunctionsRuntimeBundle, error: Err
 }
 
 function logDebug(msg: string, data?: any): void {
-  new EmulatorLog("DEBUG", "runtime-status", msg, data).log();
+  new EmulatorLog("DEBUG", "runtime-status", `[${process.pid}] ${msg}`, data).log();
 }
 
 async function invokeTrigger(
@@ -973,7 +979,7 @@ async function invokeTrigger(
   }).log();
 
   const trigger = triggers[frb.triggerId];
-  logDebug("", trigger.definition);
+  logDebug("triggerDefinition", trigger.definition);
   const mode = trigger.definition.httpsTrigger ? "HTTPS" : "BACKGROUND";
 
   logDebug(`Running ${frb.triggerId} in mode ${mode}`);